Refactor: Move common UI components to the :ui module
I moved the following common classes and files from :ui:main to the :ui module to prepare for future module separation of playintegrity and keyattestation packages:

- `InfoItem.kt`
- `InfoItemFormatter.kt`
- `ProgressConstants.kt` (renamed from `PlayIntegrityProgressConstants.kt`)
- `Base64Utils.kt`

I also updated all references to these files to point to their new locations.
