+++ android/ui/src/main/java/dev/keiji/deviceintegrity/ui/common/InfoItem.kt
@@ -1,4 +1,4 @@
-package dev.keiji.deviceintegrity.ui.main
+package dev.keiji.deviceintegrity.ui.common
 
 data class InfoItem(
     val label: String,
